cross-platform: MSVC on Win
The operator[] is further changed to satisfy the requirements of SIMD intrinsics.

isize and usize are 64-bits again.

+++ include/waffle/arithmetics.hpp
@@ -5,8 +5,8 @@
 namespace waffle
 {
 
-	using isize = std::int32_t;
-	using usize = std::uint32_t;
+	using isize = std::int64_t;
+	using usize = std::uint64_t;
 
 	using byte = std::uint8_t;
